Also fixes a number of compiler warnings, and re-instates
ParseCallback::absAddr, which should not have been deleted,
and removes ParseCallback::loadAddr which was unused and
redundant with absAddr.

Conflicts:
	dyninstAPI/src/Parsing.h
	patchAPI/src/PatchBlock.C

+++ dyninstAPI/src/process.C
@@ -4288,7 +4288,7 @@ bool process::getDeadCode
          bit != deletableBlocks.end();
          bit++) 
     {
-        if (bit->first->containingFuncs() == bit->second.size()) {
+        if ((unsigned int)bit->first->containingFuncs() == bit->second.size()) {
             delBlocks.insert(bit->first);
         }
     }
